rivos: simplify CI with a single container level
Rather than running the builds inside a distro container which is
inside the CI container (Docker in Docker mode), go back to a single
container level by asking the CI to use directly the intended container.

Use the gitlab CI script template from the rivos-sdk and leverage it for the packaging.

Additional notable functional changes:
- no longer export the build container as mostly nobody is using it.