locking/atomic: cmpxchg: Make __generic_cmpxchg_local compare against zero-extended 'old' value
__generic_cmpxchg_local takes unsigned long old/new arguments which might end up being up-cast
from smaller signed types (which will sign-extend).  The loaded compare value must be compared
against a truncated smaller type, so down-cast appropriately for each size.

The issue is apparent on 64-bit machines with code such as atomic_dec_unless_positive() will
use signedness of ints in anger.

64-bit machines generally don't use the generic cmpxchg, but development/early ports might
make use of it so make it correct.

Signed-off-by: Matt Evans <mev@rivosinc.com>