RIVOS: DCE: check PASID on kernel job submission
Avoids the case where the FD is cloned and there would be a need to
support multiple pasids for the same queue

Returns error on submits from the cloned process. Hopefully, leading to
closing of the fd to avoid delaying the release in the first process.

Also clean up some helper functions