rivos: Add TEST_ONLY and TEST_KASAN env variables
No need to trigger kasan builds/tests at every submission, let the
developer ask for it, this will reduce the CI duration (even if jobs are
run in parallel, kasan inline is twice as long as other jobs to build).

TEST_ONLY can be used to only test the test infrastructure to avoid
wasting time building the kernels.

Signed-off-by: Alexandre Ghiti <alexghiti@rivosinc.com>