RIVOS: dpa: Support for subscribing to and waiting on signals
Adds Daffy support for subscribing to signals. The 'wait_signal'
IOCTL now subscribes to a given signal, then sleeps until DUC firmware
modifies the signal (or until timeout). Notification of signal
modification is achieved via a DUC-to-host Daffy packet. To
accomplish this, this patch adds initial support for processing of the
Daffy d_queue via a dedicated workqueue.

Signed-off-by: Sean Walsh <sean@rivosinc.com>