RIVOS: DPA: Remove device memory management
Since CPU and DPA have a unified address space, We're going to model HBM
as a separate NUMA node and manage it using existing memory management
APIs. For userspace, this means allocating HBM is now a matter of mmap()
and mbind(). This allows us to effectively remove all buffer management
from the driver.

Note that we'll eventually use hmem for HBM so that it doesn't
automatically end up in the kernel page pool at boot, but for now it's
just an ordinary (CPU-less) NUMA node.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>