riscv: mm: Add memory hot add/remove support
>From an arch perspective, a couple of callbacks needs to be
implemented to support hotplugging:

arch_add_memory() This callback is responsible for updating the
linear/direct map, and call into the memory hotplugging generic code
via __add_pages().

arch_remove_memory() In this callback the linear/direct map is tore
down.

vmemmap_free() The function tears down the vmemmap mappings (if
CONFIG_SPARSEMEM_VMEMMAP is in-use), and also deallocates the backing
vmemmap pages. Note that for persistent memory, an alternative
allocator for the backing pages can be used -- the vmem_altmap. This
means that when the backing pages are cleared, extra care is needed so
that the correct deallocation method is used. Note that RISC-V
populates the vmemmap using vmemmap_populate_basepages(), so currently
no hugepages are used for the backing store.

The page table unmap/teardown functions are heavily based (copied!)
from the x86 tree. The same remove_pgd_mapping() is used in both
vmemmap_free() and arch_remove_memory(), but in the latter function
the backing pages are not removed.

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>