RIVOS: isbdm: Add random element to GID
One problem in attempting to connect 3 VMs together over ISBDM is that
they'll come up with potentially identical GIDs, making routing
impossible. Since ISBDM doesn't have a unique device identifier, get a
random 32-bit value. Stick that in bits 63:32 of the GID so that each
device (probably) ends up with a unique GID in the cluster.

Signed-off-by: Evan Green <evan@rivosinc.com>