RIVOS: isbdm: Remove WQE wr_status
The wr_status field of the isbdm_wqe was a vestige of the siw driver. In
that driver, a WQE was "activated", and then underwent several
asynchronous steps of network transfers before completing. The ISBDM
driver however can enqueue hardware commands without multiple
asynchronous steps. So maintaining a processing state machine is just
unnecessary complexity.

In most cases that were using wr_status, there was only ever actually
one possible state given the code flow within ISBDM. These were easy
cases to de-conditionalize an if statement, or remove it entirely.

One area that did need attention is that the wr_status was used as a
flag to indicate whether sending was already in progress, or needed to
be started. I replaced this with a simpler send_pending flag, which is
more clearly protected by the sq_lock. I think there may have actually
been bugs in the previous handling related to kernel and user requests
being sent at the same time, so this logic should be easier to follow.

Removing the wr_state handling also made me realize we lacked the proper
TX halt logic, should the QP error out during send completion. Add a
tx_halted flag (corresponding to the tx_suspend flag in the siw driver)
that halts TX processing regardless of the QP state.

Signed-off-by: Evan Green <evan@rivosinc.com>