RIVOS: isbdm: Remove tx_wqe() and rx_rqe()
Rather than have functions reach into shared state within the QP, pass
the WQE pointer in directly to the couple of functions that need it.
This makes the data sharing easier to follow, as there's now only a
single place where qp.{tx,rx}_ctx.wqe_active are used. This could now
easily be converted to a stack local if multiple functions needed to be
in here at once, though I didn't opt to do that to conserve stack space.

Signed-off-by: Evan Green <evan@rivosinc.com>