RIVOS: isbdm: Add solicited events support
Add a flags field to the header, and send the SOLICITED bit across the
send, so the receiver can differentiate between regular CQ entries and
solicited ones. This is basically just an "important" bit as specified
in the libibverbs API. Most of this was already implemented, we just
needed to send the bit across the wire.

Signed-off-by: Evan Green <evan@rivosinc.com>