RIVOS: isbdm: Fix RDMA transfers >2GB
isbdm_check_sgl_tx() was attempting to return a signed int, which when
negative represented an error, and when positive represented a length.
This makes RDMA transfers with a total length >=2GB look like errors.
Since the positive return value is just used to fill in wqe->bytes, and
wqe is passed into isbdm_check_sgl_tx(), just fill that in directly as a
side effect (after checking that it will fit in the u32). The function
then goes back to the more traditional convention of 0 for success and
negative for failure.

Signed-off-by: Evan Green <evan@rivosinc.com>