RIVOS: isbdm: Add loopback AMO support
Implement support for atomic operations (fetch'n'add, compare/exchange)
when targeting oneself over ISBDM. These are somewhat painful to
implement since we need a pointer to a buffer in another process to do
the atomic operation on. Even the "local" iova may not be the current
process, as WQEs can get run from a kernel worker thread.

Signed-off-by: Evan Green <evan@rivosinc.com>