RIVOS: dpa: Make daffy queue submission thread-safe
Once a packet is in the h2d queue, we cannot reuse its slot in the ring
buffer until the interrupt handler has copied out the packet and woken up
the thread (if any) waiting for its completion. Furthermore, we should
block when trying to enqueue a packet in a full h2d queue.

Fix these issues by keeping packets waiting for completion in a list, and
putting threads waiting for a slot in the queue in a waitqueue. Put all
of this in daffy_submit_sync() to reduce code duplication.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>