RIVOS: isbdm: Tighten up GID and routing
Our distributed UCC sample script was falling over with a group size of
3 machines. This is because the subnet_prefix (top 64 bits) of the
device's GID was fixed to zero, so UCC thought all devices were
connected to the same subnet. It was then sending packets to the wrong
machine, which either got dropped due to that QP number not being set up
on the other side, or absorbed down to UCX, who got confused to be
receiving packets (connection requests or responses usually) it wasn't
expecting.

To start with, let's add some checking in the kernel by sending the
intended destination lid across the wire. The reciever can then check
this field, and reject packets not destined for it. This prevents UCX
from asserting, but still doesn't actually solve the problem.

Let's also create a handshake packet that ISBDM devices send to each
other when they're initially connected. This will likely be needed for
other reasons in the future as well, but in this case we'll use it to 1)
exchange interface_ids (the lower 64-bits of the peer's GID), and 2)
negotiate a shared randomly generated subnet prefix. With this info, the
ISBDM provider can both fill out the grh properly, and present userspace
with a different subnet prefix for each ISBDM link.

Signed-off-by: Evan Green <evan@rivosinc.com>