RIVOS: isbdm: Fix list corruption during disconnect
As part of its duties, isbdm_abort_rx() clears out the
rx_ring.wait_list, which is the set of packets waiting to make it into
hardware descriptors. It iterates over the list, calling put_buf() on
the list. put_buf() will add the buffer back into the free list. This
corrupts the wait list, as the wait list head now points to entries that
are actually on the free list.

Refactor a bit to call free_buf_list() in isbdm_abort_rx() and
isbdm_abort_tx(). Add the INITIALIZE_LIST_HEAD() in free_buf_list() so
that if the device is reconnected, the wait_list is still valid (but
empty). There is a bit of a change now as we're calling free_buf()
instead of put_buf(), but the code reuse is nice, and disconnect should
be a rare enough event that the perf impact is negligible.

Signed-off-by: Evan Green <evan@rivosinc.com>