RIVOS: ci: Skip unnecessary CI jobs when run downstream of DPA
The DPA CI pipeline needs to be able to build Linux with the
rivos-defconfig for use in downstream runs of FS-Sim. Other steps
(such as running tests or building other configs) aren't needed by
the DPA and add time to the pipeline.

This commit adds rules that check if the trigger method was a
multi-project pipeline. If so, the above CI jobs not needed by the
DPA are skipped, and a downstream run of FS-Sim is triggered.

Signed-off-by: Sean Walsh <sean@rivosinc.com>