RIVOS: rivos_defconfig: Disable CONFIG_KVM
KVM is incompatible with the RASD support logic in the ISBDM driver,
as both drivers want the non-shared Guest External Interrupt (12).
Additionally, we're planning to expose H in the ISA string on GA0,
despite the fact that it does not fully support H. Disabling KVM
effectively prevents access to the unverified portions of the H
implementation.

Signed-off-by: Evan Green <evan@rivosinc.com>