RIVOS: dpa: Avoid taking p->lock in wait_signal
Use the p->signal_lock spinlock to protect signal_pages and the waiters
list. We still use p->lock to serialize calls to register_signal_pages.

We can also remove the waking up of waiting threads in remove_signal_pages
since this is only called when the fd is closed, at which point there
can be no threads in the wait_signal ioctl.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>