RIVOS: isbdm: Add RASD host commands
Add awareness of RASD host commands. Also, move the initialization of
the notify area from the RMB table init to the command ring init. In
RASD mode, we don't bother to initialize the RMB table, but we still ned
the notify area set up. Finally, add an ioctl for reading the EP_STATUS
register, so the test app can figure out if a given ISBDM is configured
in RASD mode or not.

Signed-off-by: Evan Green <evan@rivosinc.com>