RIVOS: dpa: Make unregistering signal pages implicit
There's no need to make the host unregister signal pages explicitly on
process exit. Also rename {register -> set}_signal_pages in the uAPI
and Daffy interfaces since this is a one way operation. (And if we do
need to support modifying a process's signal pages after the fact, we
can still use 'set' for this purpose.)

The forthcoming notification queue will follow a similar convention.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>