RIVOS: isbdm: Set TX ND flag if ring is full
In the hybrid-sim case, I ran into a situation where the kernel got
wedged trying to send a big packet (17 descriptors) on a small ring (16
descriptors). In this case, the whole ring is filled with descriptors
that don't have the ND bit set. We never get the TXDONE interrupt, and
never continue sending the remainder of the packet.

Set the ND bit both when it's the Last Segment in the packet, as well as
if this segment is about to make the ring completely full. This case
should basically never happen in a reasonably sized ring, but let's do
it anyway for completeness sake.

Signed-off-by: Evan Green <evan@rivosinc.com>