RIVOS: isbdm: Add ioctl to control in-memory queuing
There are now two ways to submit an RDMA command: through the in-memory
head/tail index (brought to you by RASD), and the normal way of updating
MMIO head/tail registers. Add an ioctl to allow the test program to
control which method we use, allowing us to more fully exercise the
hardware.

Signed-off-by: Evan Green <evan@rivosinc.com>