riscv: Improve PTDUMP to show RSW with non-zero value
RSW field can be used to encode 2 bits of software defined
information, currently PTDUMP only prints RSW when its value
is 1 or 3.

To fix this issue and enhance the debug experience with PTDUMP,
we use _PAGE_SOFT as the RSW mask and redefine _PAGE_SPECIAL to
(1 << 8), allow it to print the RSW with any non-zero value,
otherwise, it will print an empty string for each row.

This patch also removes the val from the struct prot_bits as
it is no longer needed.

Signed-off-by: Yu Chien Peter Lin <peterlin@andestech.com>