RIVOS: dpa: Define KILL_PASID command
The host sends the KILL_PASID command to request that the DUC terminate
any work the process has outstanding. If there is work that needs to be
forcibly terminated, the DUC responds with IN_PROGRESS and sends a
KILL_PASID message back when the kill is complete. Otherwise, the DUC
responds immediately with SUCCESS.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>