RIVOS: Add a magicbox driver
This driver allows to use the magicbox "device" to output meaningful data
from simulation. This drivers exposes the following files in
/sys/kernel/debug/magicbox:

- exit: Write only file to output an exit code from simulation
- dump_checkpoint: Dump checkpoint upon writing to this file
- dump_stat: Dump statistics upon writing this file

For instance, exit can be triggered with:

$ echo 1 > /sys/kernel/debug/magicbox/exit

Upon panic, the kernel will output 0xDEADBEEF as exit code. On reboot,
it will output 0xDEAD000 ored with the action reported by the kernel.

Some device tree node such as the following should be added to probe
this driver.

magicbox: magicbox@42000000 {
	compatible = "rivos,magicbox";
	regs = <0x0 0x42000000 0x0 0x100>;
};

Building this driver should be done using CONFIG_RIVOS_MAGICBOX.

Signed-off-by: Clément Léger <cleger@rivosinc.com>