RIVOS: isbdm: Fix rmb alloc infinite spin
The RMB allocation code was meant to make one complete rotation around
the ring, starting from where it last found an RMB, looking for a free
slot. Due to the ordering of the increment, rollover, and exit check, it
was possible for a starting position of 0 to iterate forever. Fix this
by moving the increment.

Signed-off-by: Evan Green <evan@rivosinc.com>