RIVOS: HACK: pci: Avoid DOE aborts on reset
With the new Darjeeling integration, the RoT's DOE mailboxes aren't yet
capable of dealing with abort interrupts. Linux pushes an abort through
as the first thing it does to a DOE mailbox, in order to verify liveness
and put the mailbox into a known state. This doesn't work, so the
mailbox doesn't come up. ISBDM needs the mailbox to communicate link
status changes to the RoT.

Temporarily skip the initial abort on DOE mailboxes until the hubris
implementation catches up. Follow this bug for progress:

https://rivosinc.atlassian.net/browse/SW-5251

Signed-off-by: Evan Green <evan@rivosinc.com>