RIVOS: DCE: Fix race on kernel queue release
The release process frees dma coherent memory allocations towards the
end of the process and zeroes some of the cached data.

Before this patch, this happened before the queue type was changed to
disabled. Adding another state to the queue as retiring as to not
disturn the 'delicate' semantics of the driver.

While the whole thing is a bit of a mess and need serieous rework, this
changes makes the longstanding bad access, due to the clean up process
cleaning up a queue where the memory has already been unallocated,
disappear.