rivos: create-next: Add option to exclude internals
To work with some partners, it is necessary to be able to produce a tag
that can be shared without leaking details of IP blocks that haven't yet
been shared. Start by filtering out the DPA, DCE, and ISBDM branches and
using upstream as a base.

TEST=generate a branch with `create-next` before and after the change
`git diff` to make sure they're the same.
`create-next -e` and check that internal branches aren't included.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>