misc: qemu-edu: Fix cleanup on misc_register failure
The code was missing a call to disable SVA, which would cause subsequent
probe attempts to fail.

This can be observed when there are multiple qemu-edu devices in the
system. misc_register fails when this is the case since there can only
be one /dev/qemu-edu. Unbinding the driver from the first probed device
instance and rebinding it to the second one would still fail due to the
missing SVA disabling. It works as expected after this fix.

Signed-off-by: Mattias Nissler <mnissler@rivosinc.com>