riscv: add support for SBI Supervisor Software Events extension
The SBI SSE extension allows the supervisor software to be notified by
the SBI of specific events that are not maskable. The context switch is
handled partially by the firmware which will save registers a0, a6, a7.
When entering kernel we can rely on these 3 registers to setup the stack
and save all the registers.

Since each events (and each CPU for local events) have their own
context and can preempt each other, allocate a stack for each of them
(and for each cpu for local events).

When completing the event, if we were coming from kernel with interrupts
disabled, simply return there. If coming from userspace or kernel with
interrupts enabled, simulate an interrupt exception to allow delivery of
signal to user task. For instance this can happen, when a RAS event has
been generated and a SIGBUS has been sent to a task.

Signed-off-by: Clément Léger <cleger@rivosinc.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>