iommu/riscv: Fix ATS invalidation span
The ATS invalidation format can encode a power-of-2 invalidation
size from a size-aligned base address.  The previous code led to
missing invalidations where an input range straddled a Po2 boundary
larger than the range size.

For example, an input range of 7ffff7ebe000-7ffff7ece000 should be
rounded to a 512K span of 7ffff7e80000-7ffff7f00000 (and not a 128K
span from 7ffff7ea0000 as was previously calculated from the range
alone).

Signed-off-by: Matt Evans <mev@rivosinc.com>